Add Basic Metering Report to Workflow Engine (#16398)
* Add Basic Metering Report to Workflow Engine

This commit adds a metering report struct to the workflow engine such that each step can be individually added to the
report. Metering units and values are generally treated as string values, but post values are verified to be valid
numeric.

The metering report can provide a calculated median value for all posted units.

* update mathematical operations on metering spend value

* fix tests and greater than comparison